adicionado comentarios e corrigido o tipo da variavel

+++ questao_4.py
@@ -1,3 +1,5 @@
-temp=int(input("digite a temperuta em celsius:"))
+#Escreva um algoritmo que solicite a temperatura em graus celcius
+#E converta para Fahrenheit
+temp=float(input("digite a temperuta em celsius:"))
 f=temp*1.8+32
 print(f"A a temperatura informada em Fahrenheit é:{f} F°")
